+++ src/Converter.php
@@ -50,6 +50,8 @@ class Converter implements ConverterInterface
      *
      * @param string $commonMark
      *
+     * @throws \RuntimeException
+     *
      * @return string
      *
      * @api
@@ -67,6 +69,8 @@ class Converter implements ConverterInterface
      * @see Converter::convertToHtml
      *
      * @param string $commonMark
+     *
+     * @throws \RuntimeException
      *
      * @return string
      */